% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place.R
\name{download_place}
\alias{download_place}
\title{download_place}
\usage{
download_place(place_code, loc, folder, step = 30)
}
\arguments{
\item{place_code}{A number identifying the place to be downloaded, its passed to filenames of images}

\item{loc}{vector c(lat,lng)}

\item{folder}{Defaultly it is current working directory}

\item{step}{Change of angle between two images in degrees}
}
\value{
Returnes nothing.
}
\description{
Downloads 360 degrees panorama sequence of images with defined change in angle of view
}
