% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{download_track}
\alias{download_track}
\title{download_track}
\usage{
download_track(start, end, track_code, folder = getwd(), pace = 20,
  fineness = 5, map = 1, adjust = TRUE)
}
\arguments{
\item{start}{vector c(lat,lng)}

\item{end}{vector c(lat,lng)}

\item{track_code}{A number identifying the track to be downloaded, its passed to filenames of images}

\item{folder}{Defaultly it is current working directory}

\item{pace}{Number of metres between coordinates used to download images}

\item{fineness}{number of images with adjusted heading before curve}

\item{map}{(map == 0) => downloads only images without map;  (map == 1) =>  downloads images and map of their locations; (map == 2) => downloads only summary map}

\item{adjust}{If TRUE script tryes to adjust headings in curves}

\item{pace}{metres between images}
}
\value{
Returnes vector of deviances between calculated and real location of downloaded images. Values bellow 70 are considered a sufficient match.
}
\description{
Downloads two sequances of images from Google Streetview between given coordinates, one in each direction and also can create summary maps.
}
\details{
This function first finds the route between two corrdinates using Google Directions api, then it calculates positions of images to be downloaded alongside the route so they are in predefined spacing.
afterwards it downloads the images and creates summary map. In the map the bigger red marks with letters stand for calculated positions of images and the smaller blue marks, that are connected with green line
 to the red marks, are positions of the actualy downloaded images.
 @
}
\examples{
download_track(c(50.080266, 14.447034), c(50.081416, 14.447790), 1)
download_track(c(50.064281, 14.509821), c(50.065542, 14.507479), 2, pace = 30, map = 2)
download_track(c(50.065476, 14.512228), c(50.065001, 14.514193), 3, adjust = TRUE, fineness = 7)
}
